refactors image model into a plugin
