+++ src/plugins/image/components/empty-image-prompt.scss
@@ -1,4 +1,4 @@
-@import "../../vars.sass";
+@import "../../../components/vars.sass";
 
 .image-prompt-container {
   position: absolute;
