+++ src/plugins/image/components/empty-image-prompt.tsx
