+++ src/plugins/image/components/image-tool.sass
@@ -1,4 +1,4 @@
-@import ../vars
+@import ../../../components/vars
 
 .image-tool
   width: 100%
@@ -10,7 +10,7 @@
     border: 1px solid #cccccc
 
   .loading-spinner
-    background-image: url("../../assets/Spinner-1s-200px.svg")
+    background-image: url("../../../assets/Spinner-1s-200px.svg")
     background-size: auto
     background-repeat: no-repeat
     width: 200px
