+++ src/plugins/image/components/image-tool.tsx
@@ -3,21 +3,21 @@ import React from "react";
 import ResizeObserver from "resize-observer-polyfill";
 import { observer, inject } from "mobx-react";
 import { debounce } from "lodash";
-import { BaseComponent } from "../base";
-import { EmptyImagePrompt } from "./image/empty-image-prompt";
-import { ImageToolbar } from "./image/image-toolbar";
+import { BaseComponent } from "../../../components/base";
+import { EmptyImagePrompt } from "./empty-image-prompt";
+import { ImageToolbar } from "./image-toolbar";
 import { ImageComponent } from "./image-component";
-import { IToolApi, TileResizeEntry } from "./tool-api";
-import { IToolTileProps } from "./tool-tile";
-import { IDocumentContext } from "../../models/document/document-types";
-import { debouncedSelectTile } from "../../models/stores/ui";
-import { gImageMap, ImageMapEntryType } from "../../models/image-map";
-import { ImageContentModelType } from "../../models/tools/image/image-content";
-import { ITileExportOptions } from "../../models/tools/tool-content-info";
-import { hasSelectionModifier } from "../../utilities/event-utils";
-import { ImageDragDrop } from "../utilities/image-drag-drop";
-import { isPlaceholderImage } from "../../utilities/image-utils";
-import placeholderImage from "../../assets/image_placeholder.png";
+import { IToolApi, TileResizeEntry } from "../../../components/tools/tool-api";
+import { IToolTileProps } from "../../../components/tools/tool-tile";
+import { IDocumentContext } from "../../../models/document/document-types";
+import { debouncedSelectTile } from "../../../models/stores/ui";
+import { gImageMap, ImageMapEntryType } from "../models/image-map";
+import { ImageContentModelType } from "../models/image-content";
+import { ITileExportOptions } from "../../../models/tools/tool-content-info";
+import { hasSelectionModifier } from "../../../utilities/event-utils";
+import { ImageDragDrop } from "../../../components/utilities/image-drag-drop";
+import { isPlaceholderImage } from "../utilities/image-utils";
+import placeholderImage from "../../../assets/image_placeholder.png";
 
 import "./image-tool.sass";
