+++ src/plugins/image/components/image-toolbar.scss
@@ -1,4 +1,4 @@
-@import "../../vars.sass";
+@import "../../../components/vars.sass";
 
 .image-toolbar {
   position: absolute;
