+++ src/plugins/image/models/image-content.test.ts
