+++ src/plugins/image/models/image-content.ts
@@ -1,8 +1,8 @@
 import { types, Instance, SnapshotOut } from "mobx-state-tree";
 import { exportImageTileSpec, isLegacyImageTileImport, convertLegacyImageTile } from "./image-import-export";
-import { ITileExportOptions, IDefaultContentOptions } from "../tool-content-info";
-import { ToolContentModel } from "../tool-types";
-import { isPlaceholderImage } from "../../../utilities/image-utils";
+import { ITileExportOptions, IDefaultContentOptions } from "../../../models/tools/tool-content-info";
+import { ToolContentModel } from "../../../models/tools/tool-types";
+import { isPlaceholderImage } from "../utilities/image-utils";
 import placeholderImage from "../../../assets/image_placeholder.png";
 
 export const kImageToolID = "Image";
