+++ src/plugins/image/models/image.ts
