+++ src/plugins/image/utilities/image-utils.ts
@@ -1,8 +1,7 @@
-import { DB } from "../lib/db";
+import { DB } from "../../../lib/db";
 import { ImageModelType, ImageModel } from "../models/image";
-import { ImageContentSnapshotOutType } from "../models/tools/image/image-content";
-import placeholderImage from "../assets/image_placeholder.png";
-import orgPlaceholderImage from "../assets/image_placeholder_org.png";
+import placeholderImage from "../../../assets/image_placeholder.png";
+import orgPlaceholderImage from "../../../assets/image_placeholder_org.png";
 
 const ImageConstants = {
   maxWidth: 512,
