+++ src/register-tools.ts
@@ -4,14 +4,14 @@ import "./models/tools/placeholder/placeholder-registration";
 
 const gToolRegistration: Record<string, () => void> = {
   "Diagram": () => {
-    // Currently the Diagram Tool is the only tool using SharedVariables, so we register 
+    // Currently the Diagram Tool is the only tool using SharedVariables, so we register
     // SharedVariables whenever a there is a Diagram tool in the configuration
     import(/* webpackChunkName: "Diagram" */"./plugins/diagram-viewer/diagram-registration");
     import(/* webpackChunkName: "SharedVariables" */"./plugins/shared-variables/shared-variables-registration");
   },
   "Drawing": () => import(/* webpackChunkName: "Drawing" */"./plugins/drawing-tool/drawing-registration"),
   "Geometry": () => import(/* webpackChunkName: "Geometry" */"./models/tools/geometry/geometry-registration"),
-  "Image": () => import(/* webpackChunkName: "Image" */"./models/tools/image/image-registration"),
+  "Image": () => import(/* webpackChunkName: "Image" */"./plugins/image/image-registration"),
   "Table": () => import(/* webpackChunkName: "Table" */"./models/tools/table/table-registration"),
   "Text": () => import(/* webpackChunkName: "Text" */"./models/tools/text/text-registration")
 };
